example_longhealth/requirements.txt
<<<<<<< HEAD
numpy>=1.21
PyYAML>=5.4 # For loading config
python-dotenv>=0.19.0 

# Add libraries for LLM judges you intend to support in llm_judge.py
openai>=1.0
# anthropic>=0.20
# google-generativeai>=0.4.0
=======
# Dependencies specific to this example pipeline
transformers>=4.38.0
torch>=2.0
accelerate>=0.25.0
tqdm>=4.64
bitsandbytes>=0.41.3 # Keep for quantization if still needed occasionally

# Added for semantic filtering
sentence-transformers>=2.2.2
scikit-learn>=1.1

# Optional: Keep if used elsewhere
# pandas>=1.3
>>>>>>> dec0cd3 (added semantic filtering)
